Moved JsonPretty into it's own module, added more tests
